+++ marketplace/_validators/check_examples.py
@@ -40,8 +40,8 @@ BRAIN_VERBS = frozenset({
 })
 
 COMM_VERBS = frozenset({
-    "comm.send", "comm.delivered", "comm.inbox", "comm.read", "comm.reply", "comm.thread",
-    "comm.health",
+    "comm.send", "comm.delivered", "comm.inbox", "comm.unread", "comm.read",
+    "comm.mark_read", "comm.reply", "comm.thread", "comm.health", "comm.probe",
 })
 
 SCHEDULE_VERBS = frozenset({